% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReport.R
\name{createReport}
\alias{createReport}
\title{createReport}
\usage{
createReport(
  GGIRoutputdir = NULL,
  lang = "fr",
  idsep = "_",
  desiredtz = "",
  type = NULL,
  deviceName = NULL,
  quartile_thresholds = NULL,
  maskingFile = NULL
)
}
\arguments{
\item{GGIRoutputdir}{Path to GGIR output folder, e.g. C:/output_mystudy}

\item{lang}{Character, language to use fr=french, en=english}

\item{idsep}{Character, to split filename and use first segment as participant id}

\item{desiredtz}{Character, timezone from timezone data base names, see also GGIR documentation.}

\item{type}{Character to specify type of report, current options: onepage_luxsleepactcr_A4}

\item{deviceName}{Character to be used as device name, if not specified then we call it movement sensor}

\item{quartile_thresholds}{Data.frame with thresholds (rows) between the quartiles for 5 variables (columns).}

\item{maskingFile}{Character to point to csv file with dates to be masked per ID}
}
\value{
no object is returned, a pdf is saves is save in the GGIr output directory
}
\description{
createReport
}
