% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introText.R
\name{introText}
\alias{introText}
\title{introText}
\usage{
introText(GGIRoutputdir = NULL, id = NULL, lang = "fr")
}
\arguments{
\item{GGIRoutputdir}{Path to GGIR output folder, e.g. C:/output_mystudy}

\item{id}{Character, the id of the recording}

\item{lang}{Character, language to use fr=french, en=english}
}
\value{
no object is returned, text is printed with cat
}
\description{
introText
}
