% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lux_sleep_act_cr.R
\name{plot_lux_sleep_act_cr}
\alias{plot_lux_sleep_act_cr}
\title{plot_lux_sleep_act_cr}
\usage{
plot_lux_sleep_act_cr(GGIRoutputdir, id, lang = "fr", desiredtz = "")
}
\arguments{
\item{GGIRoutputdir}{Path to GGIR output folder, e.g. C:/output_mystudy}

\item{id}{Character, participant id}

\item{lang}{Character, language to use fr=french, en=english}

\item{desiredtz}{Character, timezone from timezone data base names, see also GGIR documentation.}
}
\value{
no object is returned, only a plot is generated
}
\description{
plot_lux_sleep_act_cr
}
