% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReport.R
\name{creatReport}
\alias{creatReport}
\title{creatReport}
\usage{
creatReport(
  GGIRoutputdir = NULL,
  lang = "fr",
  idsep = "_",
  desiredtz = "",
  type = NULL,
  deviceName = NULL
)
}
\arguments{
\item{GGIRoutputdir}{Path to GGIR output folder, e.g. C:/output_mystudy}

\item{lang}{Character, language to use fr=french, en=english}

\item{idsep}{Character, to split filename and use first segment as participant id}

\item{desiredtz}{Character, timezone from timezone data base names, see also GGIR documentation.}

\item{type}{Character to specify type of report, current options: onepage_luxsleepactcr_A4}

\item{deviceName}{Character to be used as device name, if not specified then we call it movement sensor}
}
\value{
no object is returned, a pdf is saves is save in the GGIr output directory
}
\description{
creatReport
}
