% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introText.R
\name{introText}
\alias{introText}
\title{introText}
\usage{
introText(GGIRoutputdir = NULL, id = NULL, lang = "fr", deviceName = NULL)
}
\arguments{
\item{GGIRoutputdir}{Path to GGIR output folder, e.g. C:/output_mystudy}

\item{id}{Character, the id of the recording}

\item{lang}{Character, language to use fr=french, en=english}

\item{deviceName}{Character with the device name to be mention in report, which defaults to movement sensor watch in the chosen language if not specified}
}
\value{
no object is returned, text is printed with cat
}
\description{
introText
}
