% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTable.R
\name{prepareTable}
\alias{prepareTable}
\title{prepareTable}
\usage{
prepareTable(GGIRoutputdir, id, lang, maskingFile)
}
\arguments{
\item{GGIRoutputdir}{Path to GGIR output folder, e.g. C:/output_mystudy}

\item{id}{Character, the id of the recording}

\item{lang}{Character, language to use fr=french, en=english}

\item{maskingFile}{Character to point to csv file with dates to be masked per ID}
}
\value{
no object is returned, text is printed with cat
}
\description{
prepareTable
}
