% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeReportTable.R
\name{writeReportTable}
\alias{writeReportTable}
\title{writeReportTable}
\usage{
writeReportTable(data, id, lang)
}
\arguments{
\item{data}{Data.frame with data to be reported}

\item{id}{Character, the id of the recording}

\item{lang}{Character, language to use fr=french, en=english}
}
\value{
no object is returned, table is added to report.
}
\description{
writeReportTable
}
